requirements.txt
# Core dependencies
celery==5.3.6
click==8.1.7
openai==1.12.0
pydantic>=2.5.2
PyYAML==6.0.1
redis>=5.0.1
newsapi-python>=0.2.7
python-dotenv>=1.0.0

# Required by the above packages
requests>=2.31.0  # Required by newsapi-python

# OpenAI SDK
openai>=1.3.7

# News Scraping
beautifulsoup4>=4.12.0
scrapy>=2.11.0

# Task Queue & Scheduling
flower>=2.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.0

# Testing
pytest>=8.3.5
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Additional dependencies
fastapi>=0.104.1
uvicorn>=0.24.0

# Email and Storage
boto3>=1.29.6
python-multipart>=0.0.6
aiosmtplib>=2.0.2
jinja2>=3.1.2

# Language Processing
langdetect>=1.0.9
deep-translator>=1.11.4
iso639-lang>=2.1.0

# Content Analysis
textblob>=0.17.1
nltk>=3.8.1
scikit-learn>=1.3.2

# Logging and Error Handling
structlog>=24.1.0
sentry-sdk>=1.39.1

# SQL and Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Stats and Analytics
statsd>=4.0.1
psutil>=5.9.6 